+++ piezas/rook.py
@@ -6,6 +6,9 @@ class Rook(Piece):
    black_str ="♖"
    white_str ="♜"
 
+   def __init__(self, color):
+        super().__init__(color)
+
    def pieza_del_mismo_color(self, row, col, board):
         #Esta funcion sirve para ver si hay otra pieza del mismo color
         pieza = board[row][col]
@@ -43,13 +46,15 @@ class Rook(Piece):
                     possibles.append((next_row, col))
                     break
             possibles.append((next_row, col))
+        # Ordenar las posiciones para asegurar el orden correcto
+        possibles.sort()
         return possibles
        
    def posiciones_horizontales(self, row, col,board):
     #Verifica los movimientos a izquierda de la torre
         possibles = []
         for next_col in range(col - 1, -1, -1):  
-            if self.pieza_del_mismo_color((row, next_col,board)):
+            if self.pieza_del_mismo_color(row, next_col,board):
                 break
             if self.captura(row,next_col,board):
                 possibles.append((row,next_col))
@@ -58,15 +63,17 @@ class Rook(Piece):
 
     #Verifica los movimientos a derecha de la torre
         for next_col in range(col + 1, 8):  
-            if self.pieza_del_mismo_color((row, next_col,board)):
+            if self.pieza_del_mismo_color(row, next_col,board):
                 break
             if self.captura(row,next_col,board):
                 possibles.append((row,next_col))
                 break
             possibles.append((row,next_col))
-        
+
+        # Ordenar las posiciones para asegurar el orden correcto
+        possibles.sort()
         return possibles
    
-       
-
+   
+